move provider interface into the manager
